# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lume - AI Assistant Platform",
    "description": "Platform for creating and managing AI assistants with RAG",
    "version": "2.0.0"
  },
  "paths": {
...
    "/knowledge-base/collections": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "List Collections",
        "description": "List all available collections.",
        "operationId": "listCollections",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Create Collection",
        "description": "Create a new knowledge base collection.",
        "operationId": "createCollection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/collections/{collection_name}": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Get Collection Config",
        "description": "Get configuration for a specific collection.",
        "operationId": "getCollectionConfig",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Update Collection",
        "description": "Update collection metadata.",
        "operationId": "updateCollection",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Delete Collection",
        "description": "Delete a collection and all associated data.",
        "operationId": "deleteCollection",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/links": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Get Links",
        "description": "Discover links from a website and mark existing ones.",
        "operationId": "getWebsiteLinks",
        "parameters": [
          {
            "name": "base_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "URL to crawl for links",
              "title": "Base Url"
            },
            "description": "URL to crawl for links"
          },
          {
            "name": "include_external",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include external domains",
              "default": false,
              "title": "Include External"
            },
            "description": "Include external domains"
          },
          {
            "name": "collection_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Collection to check for existing URLs",
              "title": "Collection Name"
            },
            "description": "Collection to check for existing URLs"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebsiteLinkInfo"
                  },
                  "title": "Response Getwebsitelinks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/upload-websites": {
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Upload Websites",
        "description": "Start background website ingestion. Poll progress via /upload-progress/{task_id}.",
        "operationId": "uploadWebsites",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/reindex": {
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Reindex Urls",
        "description": "Delete and re-ingest URLs.",
        "operationId": "reindexUrls",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReindexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/watch-urls/{collection_name}": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Watch Urls",
        "description": "Check if website contents have changed since last scrape.",
        "operationId": "watchUrls",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchUrlsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/upload-files": {
      "post": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Upload Files",
        "description": "Start background file ingestion. Poll progress via /upload-progress/{task_id}.",
        "operationId": "uploadFiles",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadFiles"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/files/{collection_name}/{filename}": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Get File",
        "description": "Serve an uploaded file.",
        "operationId": "getFile",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge-base/upload-progress/{task_id}": {
      "get": {
        "tags": [
          "knowledge-base"
        ],
        "summary": "Get Upload Progress",
        "description": "Get progress for a website or file upload task.",
        "operationId": "getUploadProgress",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_uploadFiles": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "files"
        ],
        "title": "Body_uploadFiles"
      },
      "CollectionConfigResponse": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "dense_embedding_model": {
            "type": "string",
            "title": "Dense Embedding Model"
          },
          "dense_embedding_dim": {
            "type": "integer",
            "title": "Dense Embedding Dim"
          },
          "sparse_embedding_model": {
            "type": "string",
            "title": "Sparse Embedding Model",
            "default": "bm25"
          },
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size"
          },
          "chunk_overlap": {
            "type": "integer",
            "title": "Chunk Overlap"
          },
          "distance_metric": {
            "type": "string",
            "title": "Distance Metric"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "dense_embedding_model",
          "dense_embedding_dim",
          "chunk_size",
          "chunk_overlap",
          "distance_metric",
          "created_at",
          "updated_at"
        ],
        "title": "CollectionConfigResponse",
        "description": "The full collection configuration stored in MongoDB."
      },
      "CollectionCreateRequest": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size",
            "default": 1000
          },
          "chunk_overlap": {
            "type": "integer",
            "title": "Chunk Overlap",
            "default": 100
          },
          "distance_metric": {
            "type": "string",
            "title": "Distance Metric",
            "default": "Cosine similarity"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "embedding_model"
        ],
        "title": "CollectionCreateRequest"
      },
      "CollectionListResponse": {
        "properties": {
          "collection_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Collection Names"
          }
        },
        "type": "object",
        "required": [
          "collection_names"
        ],
        "title": "CollectionListResponse"
      },
      "CollectionUpdateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "CollectionUpdateRequest"
      },
      "CompletionStatResponse": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "variant": {
            "type": "string",
            "title": "Variant",
            "default": "info"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "CompletionStatResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProgressStageResponse": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "current": {
            "type": "integer",
            "title": "Current",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "items"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current",
            "default": false
          },
          "current_item": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Item"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "ProgressStageResponse"
      },
...
      "ReindexRequest": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "urls"
        ],
        "title": "ReindexRequest"
      },
      "TaskProgressResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/ProgressStageResponse"
            },
            "type": "array",
            "title": "Stages"
          },
          "stats": {
            "items": {
              "$ref": "#/components/schemas/CompletionStatResponse"
            },
            "type": "array",
            "title": "Stats"
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "status",
          "title",
          "message"
        ],
        "title": "TaskProgressResponse"
      },
      "TaskStartedResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "message"
        ],
        "title": "TaskStartedResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatchUrlsResponse": {
        "properties": {
          "total_urls": {
            "type": "integer",
            "title": "Total Urls"
          },
          "changed_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Urls"
          },
          "unchanged_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unchanged Urls"
          },
          "changed_count": {
            "type": "integer",
            "title": "Changed Count"
          },
          "unchanged_count": {
            "type": "integer",
            "title": "Unchanged Count"
          }
        },
        "type": "object",
        "required": [
          "total_urls",
          "changed_urls",
          "unchanged_urls",
          "changed_count",
          "unchanged_count"
        ],
        "title": "WatchUrlsResponse"
      },
      "WebsiteLinkInfo": {
        "properties": {
          "href": {
            "type": "string",
            "title": "Href"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": "Untitled"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          },
          "base_domain": {
            "type": "string",
            "title": "Base Domain",
            "default": ""
          },
          "exists_in_collection": {
            "type": "boolean",
            "title": "Exists In Collection",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "href",
          "url"
        ],
        "title": "WebsiteLinkInfo",
        "description": "A link discovered during crawling."
      },
      "WebsiteUploadRequest": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "urls"
        ],
        "title": "WebsiteUploadRequest"
      }
    }
  }
}